Handle wool correctly, move some things back, rename miscutils to platformutils
