+++ core/src/main/java/tc/oc/occ/matchshare/utils/WoolUtils.java
@@ -1,21 +1,23 @@
-package tc.oc.occ.matchshare.util;
+package tc.oc.occ.matchshare.utils;
+
+import static tc.oc.pgm.util.material.ColorUtils.COLOR_UTILS;
 
 import java.util.stream.Stream;
-import org.bukkit.DyeColor;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.PlayerInventory;
-import org.bukkit.material.Wool;
 import tc.oc.pgm.api.PGM;
 import tc.oc.pgm.api.party.Competitor;
 import tc.oc.pgm.api.player.ParticipantState;
 import tc.oc.pgm.goals.Goal;
 import tc.oc.pgm.goals.GoalMatchModule;
 import tc.oc.pgm.goals.ShowOption;
-import tc.oc.pgm.util.material.Materials;
+import tc.oc.pgm.util.material.MaterialData;
+import tc.oc.pgm.util.material.MaterialMatcher;
 import tc.oc.pgm.wool.MonumentWool;
 
 public class WoolUtils {
+  public static final MaterialMatcher WOOL = MaterialMatcher.of(m -> m.name().endsWith("WOOL"));
 
   public static boolean isHoldingWool(Player target) {
     ParticipantState player = PGM.get().getMatchManager().getParticipantState(target.getUniqueId());
@@ -23,20 +25,17 @@ public class WoolUtils {
     if (player != null) {
       Competitor team = player.getParty();
       PlayerInventory inv = target.getInventory();
-      if (inv.contains(Materials.WOOL)) {
-        holdingWool = Stream.of(inv.getContents())
-            .filter(item -> item != null && item.getType() == Materials.WOOL)
-            .anyMatch(item -> WoolUtils.isEnemyWool(item, team));
-      }
+      holdingWool = Stream.of(inv.getContents())
+          .filter(item -> item != null && WOOL.matches(item.getType()))
+          .anyMatch(item -> WoolUtils.isEnemyWool(item, team));
     }
     return holdingWool;
   }
 
   public static boolean isEnemyWool(ItemStack stack, Competitor team) {
-    if (stack == null || stack.getType() != Materials.WOOL) {
+    if (stack == null || !WOOL.matches(stack.getType())) {
       return false;
     }
-    DyeColor color = ((Wool) stack.getData()).getColor();
     GoalMatchModule gmm = team.getMatch().getModule(GoalMatchModule.class);
 
     if (gmm != null) {
@@ -44,7 +43,7 @@ public class WoolUtils {
         if (goal instanceof MonumentWool wool) {
           if (wool.hasShowOption(ShowOption.STATS)
               && !wool.isPlaced()
-              && wool.getDyeColor() == color) {
+              && COLOR_UTILS.isColor(MaterialData.item(stack), wool.getDyeColor())) {
             if (wool.getOwner() == team) {
               return true;
             }
