+++ platform/platform-modern/build.gradle.kts
@@ -6,7 +6,7 @@ plugins {
 dependencies {
     implementation(project(":core"))
     implementation(project(":util"))
-    paperweight.paperDevBundle("1.21.11-R0.1-SNAPSHOT")
+    paperweight.paperDevBundle("1.21.10-R0.1-SNAPSHOT")
 }
 
 paperweight.reobfArtifactConfiguration = io.papermc.paperweight.userdev.ReobfArtifactConfiguration.MOJANG_PRODUCTION
